% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dauPortalTools-package.R
\docType{package}
\name{dauPortalTools-package}
\alias{dauPortalTools}
\alias{dauPortalTools-package}
\title{dauPortalTools: Internal Shiny Utilities for DfE Portals}
\description{
Provides tools for building Shiny applications, including:
\itemize{
\item Database helpers
\item Logging utilities
\item GOV.UK-styled UI components
}
}
\author{
\strong{Maintainer}: Ben Smith \email{ben7.smith@education.gov.uk}

Authors:
\itemize{
  \item Ben Clewarth \email{ben.clewarth@education.gov.uk}
  \item Mark Horton \email{mark.horton@education.gov.uk}
}

}
\keyword{internal}
