% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_actions.R
\name{db_delete_notice_action}
\alias{db_delete_notice_action}
\title{Delete a Tracking Date Action Record}
\usage{
db_delete_notice_action(twn_date_id)
}
\arguments{
\item{twn_date_id}{Integer. The tracking row ID to remove.}
}
\description{
Deletes a specific action date record permanently from the database.
}
