% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-admin_modules.R
\name{db_get_app_users_matrix}
\alias{db_get_app_users_matrix}
\title{Query Access Control Matrix Configuration Directory}
\usage{
db_get_app_users_matrix(conn, app_id)
}
\arguments{
\item{conn}{A DBI database connection object.}

\item{app_id}{Integer. Application context filter frame.}
}
\value{
Data frame listing assigned users, current roles, and assignees.
}
\description{
Returns full list profiles of all users who have an active role configuration profile row.
}
