% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-admin_modules.R
\name{db_get_available_roles}
\alias{db_get_available_roles}
\title{Query Reference Table Lists of System Config Roles}
\usage{
db_get_available_roles(conn)
}
\arguments{
\item{conn}{A DBI database connection object.}
}
\value{
Data frame containing available configurations under roles_config lookup.
}
\description{
Query Reference Table Lists of System Config Roles
}
