% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_actions.R
\name{db_get_date_config}
\alias{db_get_date_config}
\title{Retrieve Date Type Configurations}
\usage{
db_get_date_config()
}
\description{
Returns all configured milestone/date types from the \code{twn_date_config} table.
}
