% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_support_records_helpers.R
\name{db_get_engaged_entity_ids}
\alias{db_get_engaged_entity_ids}
\title{Get Unique Target IDs Currently Engaged inside a Provision Track}
\usage{
db_get_engaged_entity_ids(entity_type, db_get_query = utils_db_get_query)
}
\arguments{
\item{entity_type}{Character scalar. Category type layer filter (e.g., 'School', 'Trust').}
}
\description{
Get Unique Target IDs Currently Engaged inside a Provision Track
}
