% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_get_hub_lead_urns}
\alias{db_get_hub_lead_urns}
\title{Get Distinct Lead School Identifier Values Array}
\usage{
db_get_hub_lead_urns(db_get_query = utils_db_get_query)
}
\description{
Get Distinct Lead School Identifier Values Array
}
