% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_support_records_helpers.R
\name{db_get_hub_support_urns}
\alias{db_get_hub_support_urns}
\title{Get Hub Support URNs (Deprecated)}
\usage{
db_get_hub_support_urns(db_get_query = utils_db_get_query)
}
\description{
Get Hub Support URNs (Deprecated)
}
