% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_actions.R
\name{db_get_notice_actions}
\alias{db_get_notice_actions}
\title{Retrieve Date Tracking Records for a Notice}
\usage{
db_get_notice_actions(twn_id)
}
\arguments{
\item{twn_id}{Integer. The notice identifier to filter actions by.}
}
\description{
Returns all active tracking dates and actions linked to a specific Notice ID (\code{twn_id}),
matching against configurations to pull in type names.
}
