% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-message_helpers.R
\name{db_get_portal_messages}
\alias{db_get_portal_messages}
\title{Retrieve Active Portal Messages}
\usage{
db_get_portal_messages()
}
\value{
A \code{\link{data.frame}} containing active portal messages.
}
\description{
Returns all active portal messages relevant to the current application,
including global ("catch-all") messages, resolving user identifiers.
}
\details{
The current application ID is resolved using \code{\link[=utils_get_app_id]{utils_get_app_id()}}.
The database schema is resolved via \code{\link[=utils_resolve_schema]{utils_resolve_schema()}}, and the
query is executed using \code{\link[=utils_db_get_query]{utils_db_get_query()}}.
}
