% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-admin_helpers.R
\name{db_get_roles}
\alias{db_get_roles}
\title{Retrieve Available Roles}
\usage{
db_get_roles()
}
\value{
A \code{\link{data.frame}} with role configurations.
}
\description{
Returns all roles available within the application.
}
