% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_entity_helpers.R
\name{db_get_school_list}
\alias{db_get_school_list}
\title{Retrieve School List (Deprecated)}
\usage{
db_get_school_list(db_get_query = utils_db_get_query)
}
\description{
This function is deprecated. Please update your references to use
\code{db_get_search_entities("School")} instead.
}
