% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_scrape.R
\name{db_get_scraped_posts}
\alias{db_get_scraped_posts}
\title{Retrieve Scraped Post Notices}
\usage{
db_get_scraped_posts(twn_id = NULL)
}
\arguments{
\item{twn_id}{Optional Integer. If provided, filters the results to match the specified
notice identifier. Defaults to \code{NULL}.}
}
\description{
Returns records from the \code{twn_post_notices} staging table. Can optionally be
filtered by a specific \code{twn_id} using a dynamic SQL filter.
}
