% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-user.R
\name{db_get_user_by_id}
\alias{db_get_user_by_id}
\title{Query User Record by Primary Key ID}
\usage{
db_get_user_by_id(user_id)
}
\arguments{
\item{user_id}{Integer scalar. Primary key identifier.}
}
\value{
Data frame containing user details or empty data frame.
}
\description{
Query User Record by Primary Key ID
}
