% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-user.R
\name{db_get_user_by_username}
\alias{db_get_user_by_username}
\title{Query Master Users Directory Strictly by Username}
\usage{
db_get_user_by_username(username)
}
\arguments{
\item{username}{Character scalar. Legacy AD username token.}
}
\value{
A data frame containing matching user records, or an empty data frame if none found.
}
\description{
Query Master Users Directory Strictly by Username
}
