% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_hubs_helpers.R
\name{db_hubs_lookup}
\alias{db_hubs_lookup}
\title{Retrieve Master List of All Regional Hub Profiles}
\usage{
db_hubs_lookup()
}
\value{
A data.frame with columns \link{ruhb_id} and \link{ruhb_name}.
}
\description{
Pulls a clean, unique data frame of all active hub IDs and names.
}
