% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_actions.R
\name{db_insert_date_config}
\alias{db_insert_date_config}
\title{Insert a New Date Type Configuration}
\usage{
db_insert_date_config(config_data)
}
\arguments{
\item{config_data}{A named list containing \code{twn_date_name}, \code{twn_date_desc}, and \code{created_by}.}
}
\description{
Adds a new milestone type/category to the date configuration table.
}
