% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice.R
\name{db_insert_notice}
\alias{db_insert_notice}
\title{Insert a New Notice}
\usage{
db_insert_notice(notice_data)
}
\arguments{
\item{notice_data}{A named list or row of a data frame containing fields matching
the table columns. Missing metadata fields like \code{created_on} default to \code{GETDATE()}.}
}
\value{
The result of the execution (typically number of rows affected).
}
\description{
Inserts a new record into the \code{twn_all_notices} table.
}
