% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_actions.R
\name{db_insert_notice_action}
\alias{db_insert_notice_action}
\title{Insert a New Tracking Date Action}
\usage{
db_insert_notice_action(action_data)
}
\arguments{
\item{action_data}{A named list containing \code{twn_id}, \code{twn_date_type_id},
\code{twn_date}, \code{twn_date_notes}, and \code{created_by}.}
}
\description{
Adds a specific tracked date or action baseline against a notice.
}
