% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_config.R
\name{db_insert_status_config}
\alias{db_insert_status_config}
\title{Insert a New Status Configuration}
\usage{
db_insert_status_config(status_data)
}
\arguments{
\item{status_data}{A named list containing \code{twn_status_name}, \code{twn_status_desc},
and \code{created_by}.}
}
\value{
The result of the execution.
}
\description{
Adds a new status type to the lookup configuration table.
}
