% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_config.R
\name{db_insert_type_config}
\alias{db_insert_type_config}
\title{Insert a New Notice Type Configuration}
\usage{
db_insert_type_config(type_data)
}
\arguments{
\item{type_data}{A named list containing \code{twn_type_name}, \code{twn_type_desc},
\code{created_by}, \code{coc}, and \code{web_section}.}
}
\value{
The result of the execution.
}
\description{
Adds a new notice type to the lookup configuration table.
}
