% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_scrape.R
\name{db_link_post_to_master}
\alias{db_link_post_to_master}
\title{Bind a Master ID to a Scraped Staging Record}
\usage{
db_link_post_to_master(post_twn_id, twn_id)
}
\arguments{
\item{post_twn_id}{Integer. The unique row ID of the staging table.}

\item{twn_id}{Integer. The primary key ID of the master notice entry.}
}
\description{
Updates only the \code{twn_id} relationship linkage field for a staging entry.
}
