% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-rue_event_helpers.R
\name{db_ru_add_event}
\alias{db_ru_add_event}
\title{Add a Point-in-Time Interaction Log}
\usage{
db_ru_add_event(
  event_type_id,
  event_sub_variety_id = 0,
  entity_id,
  entity_type,
  event_date,
  summary_notes = NULL,
  user_id = NULL
)
}
\description{
Add a Point-in-Time Interaction Log
}
