% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-rue_event_helpers.R
\name{db_ru_add_event_type}
\alias{db_ru_add_event_type}
\title{Add a New Event Type Configuration Option}
\usage{
db_ru_add_event_type(
  name,
  description = NULL,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Add a New Event Type Configuration Option
}
