% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-rue_event_helpers.R
\name{db_ru_get_event_types}
\alias{db_ru_get_event_types}
\title{Retrieve Event Types Catalog}
\usage{
db_ru_get_event_types(ruevt_id = NULL, db_get_query = utils_db_get_query)
}
\arguments{
\item{ruevt_id}{Integer scalar or \code{NULL}. Filter for a specific event type ID.}

\item{db_get_query}{Function used to execute the query (default: \code{utils_db_get_query}).}
}
\description{
Retrieve Event Types Catalog
}
