% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_summaries.R
\name{db_ru_get_regional_summary}
\alias{db_ru_get_regional_summary}
\title{Retrieve RISE Universal Hubs Regional Summary by GOR}
\usage{
db_ru_get_regional_summary(schema = utils_resolve_schema("db_schema_01r"))
}
\arguments{
\item{schema}{Character target RISE schema (e.g. "\link{01_RISE}" or "\link{01_RISE_b}").}

\item{con}{A DBI database connection.}
}
\value{
A data.frame containing regional aggregates.
}
\description{
Executes an entity-level regional aggregation across the 9 English GORs
based on active search views and pre-aggregated transactional tables.
}
