% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-rue_event_helpers.R
\name{db_ru_save_event_action_response}
\alias{db_ru_save_event_action_response}
\title{Upsert or Commit a Ledger Response String Entry Value}
\usage{
db_ru_save_event_action_response(
  event_id,
  rueva_id,
  response_value,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Upsert or Commit a Ledger Response String Entry Value
}
