% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-rue_event_helpers.R
\name{db_ru_update_event_type}
\alias{db_ru_update_event_type}
\title{Update an Existing Event Type Catalog Description}
\usage{
db_ru_update_event_type(ruevt_id, name, description, user_id = NULL)
}
\description{
Update an Existing Event Type Catalog Description
}
