% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_actions_helpers.R
\name{db_ruh_add_action}
\alias{db_ruh_add_action}
\title{Add a Pre-Configured Action Definition (Legacy Layer)}
\usage{
db_ruh_add_action(
  hub_id,
  ruht_id,
  action_name,
  description,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Add a Pre-Configured Action Definition (Legacy Layer)
}
