% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_add_blank_lead_school}
\alias{db_ruh_add_blank_lead_school}
\title{Add a Linked Lead School Hub Center}
\usage{
db_ruh_add_blank_lead_school(
  hub_id,
  urn,
  user_id,
  date_start,
  comment = NULL,
  db_get_query = utils_db_get_query
)
}
\description{
Add a Linked Lead School Hub Center
}
