% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_add_blank_lead_support}
\alias{db_ruh_add_blank_lead_support}
\title{Add a Flexible Lead Support Assignment Instance}
\usage{
db_ruh_add_blank_lead_support(
  lead_type,
  lead_id,
  hub_id = NULL,
  event_id = NULL,
  user_id,
  date_start,
  comment = NULL,
  db_get_query = utils_db_get_query
)
}
\arguments{
\item{lead_type}{Character string defining the provider type ('School', 'Trust', etc.).}

\item{lead_id}{Integer tracking URN or ID of the provider.}

\item{hub_id}{Optional integer reference to a Hub context.}

\item{event_id}{Optional integer reference to an Event context.}

\item{user_id}{Character string identifier of the editing user.}

\item{date_start}{Character string or Date representing the start of the assignment.}

\item{comment}{Optional character string narrative comments.}

\item{db_get_query}{Function used to execute the query, defaults to utils_db_get_query.}
}
\description{
Add a Flexible Lead Support Assignment Instance
}
