% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_support_records_helpers.R
\name{db_ruh_add_blank_support_school}
\alias{db_ruh_add_blank_support_school}
\title{Add Blank Support School (Deprecated)}
\usage{
db_ruh_add_blank_support_school(
  hub_id,
  ruht_id,
  urn,
  lead_school_id = NULL,
  user_id,
  date_start,
  comment = NULL,
  db_get_query = utils_db_get_query
)
}
\description{
Add Blank Support School (Deprecated)
}
