% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_actions_helpers.R
\name{db_ruh_add_blueprint_field}
\alias{db_ruh_add_blueprint_field}
\title{Register a New Hub Blueprint Input Field Configuration}
\usage{
db_ruh_add_blueprint_field(
  ruht_id,
  ruhsc_id,
  field_name,
  description,
  rule_type,
  is_required,
  user_id
)
}
\description{
Register a New Hub Blueprint Input Field Configuration
}
