% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_hubs_helpers.R
\name{db_ruh_add_hub}
\alias{db_ruh_add_hub}
\title{Add a Primary Hub Container Profile Record}
\usage{
db_ruh_add_hub(hub_name, user_id, db_get_query = utils_db_get_query)
}
\description{
Add a Primary Hub Container Profile Record
}
