% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_actions_helpers.R
\name{db_ruh_add_sub_category}
\alias{db_ruh_add_sub_category}
\title{Register a New Hub Provision Sub-Category}
\usage{
db_ruh_add_sub_category(ruht_id, ruhb_id, name, description, user_id)
}
\description{
Register a New Hub Provision Sub-Category
}
