% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_support_types_helpers.R
\name{db_ruh_add_support_type}
\alias{db_ruh_add_support_type}
\title{Add a New Support Framework Tracking Track}
\usage{
db_ruh_add_support_type(
  hub_id,
  name,
  description = NULL,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Add a New Support Framework Tracking Track
}
