% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_actions_helpers.R
\name{db_ruh_get_actions}
\alias{db_ruh_get_actions}
\title{Retrieve Action Configuration Catalog Definitions (Legacy Layer)}
\usage{
db_ruh_get_actions(ruha_id = NULL, db_get_query = utils_db_get_query)
}
\description{
Retrieve Action Configuration Catalog Definitions (Legacy Layer)
}
