% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_get_assigned_support_records}
\alias{db_ruh_get_assigned_support_records}
\title{Get Assigned Granular Support Records (Receivers)}
\usage{
db_ruh_get_assigned_support_records(
  ruhls_id,
  db_get_query = utils_db_get_query
)
}
\description{
Get Assigned Granular Support Records (Receivers)
}
