% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_get_lead_cohorts}
\alias{db_ruh_get_lead_cohorts}
\title{Get Cohorts Assigned to a Specific Lead Support Record}
\usage{
db_ruh_get_lead_cohorts(ruhls_id, db_get_query = utils_db_get_query)
}
\description{
Get Cohorts Assigned to a Specific Lead Support Record
}
