% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_get_lead_schools}
\alias{db_ruh_get_lead_schools}
\title{Retrieve Lead School Records Catalog}
\usage{
db_ruh_get_lead_schools(
  ruhl_id = NULL,
  hub_id = NULL,
  db_get_query = utils_db_get_query
)
}
\description{
Retrieve Lead School Records Catalog
}
