% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_support_records_helpers.R
\name{db_ruh_get_support_schools}
\alias{db_ruh_get_support_schools}
\title{Get Support Schools (Deprecated)}
\usage{
db_ruh_get_support_schools(
  ruhs_id = NULL,
  hub_id = NULL,
  db_get_query = utils_db_get_query
)
}
\description{
Get Support Schools (Deprecated)
}
