% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_support_types_helpers.R
\name{db_ruh_get_support_types}
\alias{db_ruh_get_support_types}
\title{Retrieve Active Support Framework Provisioning Types}
\usage{
db_ruh_get_support_types(hub_id = NULL, db_get_query = utils_db_get_query)
}
\description{
Retrieve Active Support Framework Provisioning Types
}
