% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_set_assigned_support_records}
\alias{db_ruh_set_assigned_support_records}
\title{Assign Target Support Records to Provider Context}
\usage{
db_ruh_set_assigned_support_records(
  ruhls_id,
  ruhsr_ids,
  db_get_query = utils_db_get_query
)
}
\description{
Assign Target Support Records to Provider Context
}
