% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_set_lead_cohorts}
\alias{db_ruh_set_lead_cohorts}
\title{Set Cohorts for a Lead Support Record (Sync Ledger)}
\usage{
db_ruh_set_lead_cohorts(
  ruhls_id,
  cohort_ids,
  db_get_query = utils_db_get_query
)
}
\description{
Set Cohorts for a Lead Support Record (Sync Ledger)
}
