% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_actions_helpers.R
\name{db_ruh_update_action}
\alias{db_ruh_update_action}
\title{Update an Existing Action Catalog Item (Legacy Layer)}
\usage{
db_ruh_update_action(
  ruha_id,
  hub_id,
  ruht_id,
  action_name,
  description,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Update an Existing Action Catalog Item (Legacy Layer)
}
