% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_hubs_helpers.R
\name{db_ruh_update_hub}
\alias{db_ruh_update_hub}
\title{Update an Existing Hub Container Name Attribute}
\usage{
db_ruh_update_hub(hub_id, hub_name, user_id)
}
\description{
Update an Existing Hub Container Name Attribute
}
