% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_update_lead_school}
\alias{db_ruh_update_lead_school}
\title{Update an Existing Linked Lead School Assignment}
\usage{
db_ruh_update_lead_school(
  ruhl_id,
  hub_id,
  date_active,
  date_ended = NULL,
  is_active,
  comment = NULL,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Update an Existing Linked Lead School Assignment
}
