% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_lead_schools_helpers.R
\name{db_ruh_update_lead_support}
\alias{db_ruh_update_lead_support}
\title{Update an Existing Master Lead Support Assignment Ledger}
\usage{
db_ruh_update_lead_support(
  ruhls_id,
  lead_type,
  lead_id,
  hub_id = NULL,
  event_id = NULL,
  date_active,
  date_ended = NULL,
  is_active,
  comment = NULL,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Update an Existing Master Lead Support Assignment Ledger
}
