% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_support_records_helpers.R
\name{db_ruh_update_support_school}
\alias{db_ruh_update_support_school}
\title{Update Support School (Deprecated)}
\usage{
db_ruh_update_support_school(
  ruhs_id,
  hub_id,
  ruht_id,
  lead_school_id = NULL,
  date_active,
  date_ended = NULL,
  is_active,
  comment = NULL,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Update Support School (Deprecated)
}
