% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ruh_support_types_helpers.R
\name{db_ruh_update_support_type}
\alias{db_ruh_update_support_type}
\title{Update an Existing Support Track Configuration Blueprint}
\usage{
db_ruh_update_support_type(ruht_id, name, description = NULL, user_id)
}
\description{
Update an Existing Support Track Configuration Blueprint
}
