% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_actions.R
\name{db_update_date_config}
\alias{db_update_date_config}
\title{Update an Existing Date Type Configuration}
\usage{
db_update_date_config(date_config_id, update_data)
}
\arguments{
\item{date_config_id}{Integer. The ID of the configuration record to update.}

\item{update_data}{A named list containing \code{twn_date_name}, \code{twn_date_desc}, and \code{updated_by}.}
}
\description{
Modifies meta-details for a specific date type configuration record.
}
