% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice.R
\name{db_update_notice}
\alias{db_update_notice}
\title{Update an Existing Notice Record}
\usage{
db_update_notice(twn_id, update_data)
}
\arguments{
\item{twn_id}{Integer. The unique identifier of the notice record to update.}

\item{update_data}{A named list containing the keys and new values to update.}
}
\value{
The result of the execution.
}
\description{
Updates all modifiable fields for a specific notice identified by \code{twn_id}.
}
