% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_actions.R
\name{db_update_notice_action}
\alias{db_update_notice_action}
\title{Update an Existing Tracking Date Action}
\usage{
db_update_notice_action(twn_date_id, update_data)
}
\arguments{
\item{twn_date_id}{Integer. The primary key of the targeted tracking record row.}

\item{update_data}{A named list containing \code{twn_date_type_id}, \code{twn_date},
\code{twn_date_notes}, and \code{updated_by}.}
}
\description{
Modifies a tracked milestone row based on its unique tracking primary key (\code{twn_date_id}).
}
