% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_config.R
\name{db_update_status_config}
\alias{db_update_status_config}
\title{Update an Existing Status Configuration}
\usage{
db_update_status_config(status_id, update_data)
}
\arguments{
\item{status_id}{Integer. The \code{twn_status_id} of the record to modify.}

\item{update_data}{A named list containing \code{twn_status_name}, \code{twn_status_desc},
and \code{updated_by}.}
}
\value{
The result of the execution.
}
\description{
Updates the details of a specific status configuration record.
}
