% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-warning_notice_config.R
\name{db_update_type_config}
\alias{db_update_type_config}
\title{Update an Existing Notice Type Configuration}
\usage{
db_update_type_config(type_id, update_data)
}
\arguments{
\item{type_id}{Integer. The \code{twn_type_id} of the record to modify.}

\item{update_data}{A named list containing \code{twn_type_name}, \code{twn_type_desc},
\code{updated_by}, \code{coc}, and \code{web_section}.}
}
\value{
The result of the execution.
}
\description{
Updates the details of a specific notice type configuration record.
}
