% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-user.R
\name{db_update_user_last_login}
\alias{db_update_user_last_login}
\title{Update User Last Login Timestamp}
\usage{
db_update_user_last_login(user_id)
}
\arguments{
\item{user_id}{Integer canonical user_id.}
}
\value{
Logical indicating whether a row was affected.
}
\description{
Sets the last_login column to the current UTC timestamp for a specific user_id.
}
