% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_allocations.R
\name{delete_ru_lead_allocation}
\alias{delete_ru_lead_allocation}
\title{Hard delete a lead allocation record}
\usage{
delete_ru_lead_allocation(rula_id)
}
\arguments{
\item{rula_id}{Integer. The primary key ID to delete.}
}
\value{
Integer number of affected rows, or NULL on error.
}
\description{
Hard delete a lead allocation record
}
