% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_schools.R
\name{delete_ru_lead_school}
\alias{delete_ru_lead_school}
\title{Hard delete a lead school record}
\usage{
delete_ru_lead_school(ruhl_id)
}
\arguments{
\item{ruhl_id}{Integer. The primary key ID to delete.}
}
\value{
Integer number of affected rows, or NULL on error.
}
\description{
Hard delete a lead school record
}
