% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-log_event.R
\docType{data}
\name{.log_con}
\alias{.log_con}
\title{Log an Event Message}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
.log_con
}
\arguments{
\item{message}{Character string containing the message to log.}
}
\value{
Invisibly returns NULL.
}
\description{
Logs a timestamped message to a file and optionally to the console.
}
\keyword{datasets}
