% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-render_entity_overview.R
\name{entity_render_overview}
\alias{entity_render_overview}
\title{Render Parent Entity Overview Panel with Full Diagnostic Guard Rails}
\usage{
entity_render_overview(
  entity_type = c("trust", "la", "diocese"),
  entity_id,
  id = paste0("entity_overview_", entity_type, "_", entity_id)
)
}
\description{
Render Parent Entity Overview Panel with Full Diagnostic Guard Rails
}
