% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-user_helpers.R
\name{get_user_id}
\alias{get_user_id}
\title{Old User Identity Getter (Deprecated)}
\usage{
get_user_id(login_token)
}
\description{
This function is deprecated. Please shift your application workflow
to track the strict integer \code{user_id} using \code{db_user_create()}.
}
