% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-user_helpers.R
\name{get_user_role_legacy}
\alias{get_user_role_legacy}
\title{Old Legacy User Role Lookup (Deprecated)}
\usage{
get_user_role_legacy(login_token)
}
\description{
Re-routes old token queries using character names into the
new normalized integer architecture.
}
