% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generate_url.R
\name{gias_school_url}
\alias{gias_school_url}
\title{Generate GIAS School URL}
\usage{
gias_school_url(urn)
}
\arguments{
\item{urn}{Character scalar. Unique Reference Number (URN) identifying a school.}
}
\value{
Character scalar containing the GIAS school URL, or \code{NULL} if the
input is missing or invalid.
}
\description{
Generate the Get Information About Schools (GIAS) page for a given
school URN.
}
\examples{
gias_school_url("123456")
gias_school_url("")

}
