% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{input_to_date}
\alias{input_to_date}
\title{Combine GOV date inputs back into Date}
\usage{
input_to_date(field, input)
}
\arguments{
\item{field}{Field base name}

\item{input}{Shiny input object}
}
\value{
Date or NA
}
\description{
Reads day/month/year inputs and returns a Date or NA
}
