% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-log.R
\name{log_summary}
\alias{log_summary}
\title{Deprecated Summary Logger}
\usage{
log_summary(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
This function is deprecated. Use \code{\link[=log_event]{log_event()}} instead.
}
