% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-make_shiny_link.R
\name{make_shiny_links}
\alias{make_shiny_links}
\title{Create Multiple Shiny HTML Links#'#' Generates a named list of Shiny HTML anchor tags from a named vector of URLs.}
\usage{
make_shiny_links(urls)
}
\arguments{
\item{urls}{Named character vector. Names represent the link text, and
values represent the URLs.}
}
\value{
A named list of \code{shiny::HTML} objects.
}
\description{
Create Multiple Shiny HTML Links#'#' Generates a named list of Shiny HTML anchor tags from a named vector of URLs.
}
\details{
Each element of \code{urls} is passed to \code{\link[=make_shiny_link]{make_shiny_link()}}. The output
preserves the original names.
}
\examples{
make_shiny_links(
  c("DfE" = "https://www.gov.uk/government/organisations/department-for-education")
)

}
\seealso{
\code{\link[=make_shiny_link]{make_shiny_link()}}
}
