% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-analytics.R
\name{record_login}
\alias{record_login}
\title{Legacy App Login Tracking Bridge (Deprecated)}
\usage{
record_login(user = "Guest")
}
\description{
Catches legacy operations calling analytic frameworks that haven't shifted to
the new structure yet.
}
