% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-ru_bi.R
\name{ru_portal_health_server}
\alias{ru_portal_health_server}
\title{RISE Universal Hubs Portal Health Server}
\usage{
ru_portal_health_server(id)
}
\arguments{
\item{id}{Character module namespace ID.}
}
\description{
RISE Universal Hubs Portal Health Server
}
