% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-ru_bi.R
\name{ru_portal_health_ui}
\alias{ru_portal_health_ui}
\title{RISE Universal Hubs Portal Health UI}
\usage{
ru_portal_health_ui(id)
}
\arguments{
\item{id}{Character module namespace ID.}
}
\value{
A Shiny tagList of bslib value boxes.
}
\description{
RISE Universal Hubs Portal Health UI
}
