% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generate_url.R
\name{scp_sc_url}
\alias{scp_sc_url}
\title{Generate Significant Change Portal URL}
\usage{
scp_sc_url(sigchange_id)
}
\arguments{
\item{sigchange_id}{Character scalar. Identifier for a significant change.}
}
\value{
Character scalar containing the portal URL, or \code{NULL} if the
input is missing or invalid.
}
\description{
Constructs a URL to the internal significant change portal for a given
significant change ID.
}
\examples{
scp_sc_url("1234")
scp_sc_url(NULL)

}
