% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_helpers.R
\name{select_ru_event_lead_providers_by_type}
\alias{select_ru_event_lead_providers_by_type}
\title{Select Lead Providers allocated to a specific Event Type}
\usage{
select_ru_event_lead_providers_by_type(ruevt_id)
}
\arguments{
\item{ruevt_id}{Integer. The event master / type ID.}
}
\value{
A \code{tibble} of lead schools allocated to events under this type.
}
\description{
Select Lead Providers allocated to a specific Event Type
}
