% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_allocations.R
\name{select_ru_lead_allocation_by_id}
\alias{select_ru_lead_allocation_by_id}
\title{Select a single lead allocation record by ID}
\usage{
select_ru_lead_allocation_by_id(rula_id)
}
\arguments{
\item{rula_id}{Integer. The primary key ID (\code{rula_id}).}
}
\value{
A single-row \code{tibble}, empty tibble, or NULL on error.
}
\description{
Select a single lead allocation record by ID
}
