% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_allocations.R
\name{select_ru_lead_allocations_all}
\alias{select_ru_lead_allocations_all}
\title{Select all lead allocation records}
\usage{
select_ru_lead_allocations_all()
}
\value{
A \code{tibble} containing all records, or NULL on error.
}
\description{
Select all lead allocation records
}
