% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_helpers.R
\name{select_ru_lead_allocations_by_lead_with_targets}
\alias{select_ru_lead_allocations_by_lead_with_targets}
\title{Select all active/historical support allocations linked to a lead school}
\usage{
select_ru_lead_allocations_by_lead_with_targets(ruhl_id)
}
\arguments{
\item{ruhl_id}{Integer. The lead school primary key.}
}
\value{
A \code{tibble} of allocations joined with target school/support contract metadata.
}
\description{
Select all active/historical support allocations linked to a lead school
}
