% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_schools.R
\name{select_ru_lead_school_by_id}
\alias{select_ru_lead_school_by_id}
\title{Select a single lead school record by primary key ID}
\usage{
select_ru_lead_school_by_id(ruhl_id)
}
\arguments{
\item{ruhl_id}{Integer. The lead school record ID (\code{ruhl_id}).}
}
\value{
A single-row \code{tibble}, empty tibble, or NULL on error.
}
\description{
Select a single lead school record by primary key ID
}
