% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_schools.R
\name{select_ru_lead_schools_all}
\alias{select_ru_lead_schools_all}
\title{Select all lead school records}
\usage{
select_ru_lead_schools_all()
}
\value{
A \code{tibble} containing all lead school records, or NULL on error.
}
\description{
Select all lead school records
}
