% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_schools.R
\name{select_ru_lead_schools_by_hub}
\alias{select_ru_lead_schools_by_hub}
\title{Select lead school records associated with a specific Hub}
\usage{
select_ru_lead_schools_by_hub(ruhb_id)
}
\arguments{
\item{ruhb_id}{Integer. The hub ID (\code{ruhb_id}).}
}
\value{
A \code{tibble} of matching lead schools, or NULL on error.
}
\description{
Select lead school records associated with a specific Hub
}
