% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_helpers.R
\name{select_ru_lead_schools_by_hub_with_metrics}
\alias{select_ru_lead_schools_by_hub_with_metrics}
\title{Select Lead Schools associated with a Hub including active caseload metrics}
\usage{
select_ru_lead_schools_by_hub_with_metrics(hub_id)
}
\arguments{
\item{hub_id}{Integer. Primary key of the Hub.}
}
\value{
A \code{tibble} containing lead schools and their caseload counts.
}
\description{
Select Lead Schools associated with a Hub including active caseload metrics
}
