% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generate_url.R
\name{slic_urn_url}
\alias{slic_urn_url}
\title{Generate SLIC Portal URLs}
\usage{
slic_urn_url(urn)
}
\arguments{
\item{urn}{Character vector (or coercible to character) of URNs.}
}
\value{
A character vector of SLIC URLs. Invalid or missing values
are returned as \code{NA_character_}.
}
\description{
Constructs URLs to the SLIC portal for one or more school URNs.
This function is vectorised and returns one URL per input value.
}
\examples{
slic_urn_url("1234")
slic_urn_url(c("1234", NA, ""))

}
