% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-ru_filter_helpers.R
\name{ui_filter_event_sub_variety_select}
\alias{ui_filter_event_sub_variety_select}
\title{Build a Cascading Event Sub-Variety Selection Inputs Wrapper}
\usage{
ui_filter_event_sub_variety_select(id, label = "Event Cohort / Sub-Variety:")
}
\description{
Build a Cascading Event Sub-Variety Selection Inputs Wrapper
}
