% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-ru_filter_helpers.R
\name{ui_filter_event_type_select}
\alias{ui_filter_event_type_select}
\title{Build a Dynamic Event Type Selection Selector}
\usage{
ui_filter_event_type_select(id, label = "Select Event Interaction Category:")
}
\arguments{
\item{id}{Character scalar.}

\item{label}{Character scalar.}
}
\description{
Build a Dynamic Event Type Selection Selector
}
