% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-quality_helpers.R
\name{ui_quality_record_table}
\alias{ui_quality_record_table}
\title{Quality Issues Table for a Specific Record}
\usage{
ui_quality_record_table(id)
}
\arguments{
\item{id}{Character scalar. Shiny module ID.}
}
\value{
A \code{shiny.tag.list} containing the targeted DataTable framework.
}
\description{
Module UI to display a clean table of unresolved quality issues for a given record.
}
