% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-quality_helpers.R
\name{ui_quality_wrapper}
\alias{ui_quality_wrapper}
\title{Quality Table Module UI (Dashboard View)}
\usage{
ui_quality_wrapper(id)
}
\arguments{
\item{id}{Character scalar. Shiny module ID.}
}
\description{
Quality Table Module UI (Dashboard View)
}
