% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-audit_helpers.R
\name{ui_render_record_metadata_banner}
\alias{ui_render_record_metadata_banner}
\title{Render Audit Record Status Metadata Banner}
\usage{
ui_render_record_metadata_banner(
  cuid = NULL,
  cdate = NULL,
  euid = NULL,
  edate = NULL,
  session = NULL
)
}
\arguments{
\item{cuid}{Integer/Character. Created user ID or email.}

\item{cdate}{Date/POSIXct/Character. Creation timestamp.}

\item{euid}{Integer/Character. Edited/Modified user ID or email.}

\item{edate}{Date/POSIXct/Character. Edited/Modified timestamp.}

\item{session}{Shiny session object (optional).}
}
\value{
A shiny.tag HTML element.
}
\description{
Takes creation and modification audit values and formats them into a
standard GDS-styled inset metadata panel.
}
