% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-user_auth.R
\name{utils_convert_email_to_ad_token}
\alias{utils_convert_email_to_ad_token}
\title{Convert Email to AD Username Heuristic}
\usage{
utils_convert_email_to_ad_token(email)
}
\arguments{
\item{email}{Character scalar.}
}
\value{
Character scalar or NULL if format cannot be parsed.
}
\description{
Handles:
\itemize{
\item Optional digits: 'ben.smith' -> 'bsmith', 'ben7.smith' -> 'bsmith7'
\item Hyphenated surnames: 'ben.smith-jones' -> 'bsmith-jones'
\item Hyphenated forenames: 'mary-jane.watson' -> 'mwatson'
\item Apostrophes: "siobhan.o'connor" -> 'soconnor'
}
}
