% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-resilience.R
\name{utils_db_validate}
\alias{utils_db_validate}
\title{Validate Active Connection Status Health}
\usage{
utils_db_validate(conn)
}
\arguments{
\item{conn}{A DBI database connection object.}
}
\value{
Logical scalar. TRUE if connection successfully responds, FALSE otherwise.
}
\description{
Verifies an existing database connection object is valid and responsive.
}
