% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-user_auth.R
\name{utils_get_user}
\alias{utils_get_user}
\title{Get Current Portal Username Token}
\usage{
utils_get_user(session = NULL, fallback = "Guest")
}
\arguments{
\item{session}{Optional Shiny session object.}

\item{fallback}{Character scalar. Defaults to \code{"Guest"}.}
}
\value{
Character scalar representing the active identity token.
}
\description{
Retrieves identity from Posit Connect headers, config override, or fallback.
}
