% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-user_auth.R
\name{utils_get_user_roles}
\alias{utils_get_user_roles}
\title{Retrieve App-Specific User Roles}
\usage{
utils_get_user_roles(user_id)
}
\arguments{
\item{user_id}{Integer scalar. Primary key identifier.}
}
\value{
Character vector of active roles.
}
\description{
Retrieve App-Specific User Roles
}
