% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-user_auth.R
\name{utils_user_has_permission}
\alias{utils_user_has_permission}
\title{Check Session Authorization Permission Flag}
\usage{
utils_user_has_permission(user_id, permission_name)
}
\arguments{
\item{user_id}{Integer scalar. Resolved user profile identifier.}

\item{permission_name}{Character scalar. Feature flag string.}
}
\value{
Logical scalar. TRUE if authorized, FALSE otherwise.
}
\description{
Check Session Authorization Permission Flag
}
