% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wn_webscrape.R
\name{wn_get_all_notice_links}
\alias{wn_get_all_notice_links}
\title{Collect All Notice Links Across Regions}
\usage{
wn_get_all_notice_links()
}
\value{
A character vector of fully qualified GOV.UK notice URLs.
}
\description{
Scrapes the regional index pages defined in the config and returns a complete,
unique character vector of individual warning notice paths.
}
