% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wn_webscrape.R
\name{wn_scrape_all_pages}
\alias{wn_scrape_all_pages}
\title{Scrape Attributes From a Vector of Notice Pages}
\usage{
wn_scrape_all_pages(notice_urls)
}
\arguments{
\item{notice_urls}{Character vector. The list of exact GOV.UK pages to crawl.}
}
\value{
A processed tibble/dataframe containing notice metadata fields.
}
\description{
Loops through a provided list of URLs, parses out standard attributes defensively,
and structures the data into a single dataframe.
}
