% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_allocations.R
\name{create_ru_lead_allocation}
\alias{create_ru_lead_allocation}
\title{Insert a new lead allocation record}
\usage{
create_ru_lead_allocation(
  ruhl_id,
  support_source_type,
  support_source_id,
  rula_dateactive = Sys.Date(),
  rula_dateended = NULL,
  rula_active = TRUE,
  rula_comment = NULL,
  created_by = NULL
)
}
\arguments{
\item{ruhl_id}{Integer. Foreign key to \code{ru_lead_schools}.}

\item{support_source_type}{Character. Support mechanism (\code{'HUB'} or \code{'EVENT'}).}

\item{support_source_id}{Integer. The ID of the hub or event.}

\item{rula_dateactive}{Date or character (\code{'YYYY-MM-DD'}).}

\item{rula_dateended}{Date or character (\code{'YYYY-MM-DD'}). Optional.}

\item{rula_active}{Logical. Default is \code{TRUE}.}

\item{rula_comment}{Character. Optional notes or context.}

\item{created_by}{Character. Username or audit identifier.}
}
\value{
Integer \code{rula_id} of the inserted record, or NULL on error.
}
\description{
Insert a new lead allocation record
}
