% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-portal_user_admin_helpers.R
\name{db_get_app_users}
\alias{db_get_app_users}
\title{Fetch app users and their role (default "User" if none)}
\usage{
db_get_app_users(conn, app_id)
}
\description{
Fetch app users and their role (default "User" if none)
}
