% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-portal_user_admin_helpers.R
\name{db_get_app_users}
\alias{db_get_app_users}
\title{Fetch application users and their assigned role}
\usage{
db_get_app_users(...)
}
\description{
Fetch application users and their assigned role
}
