% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-get_portal_messages.R
\name{db_get_portal_messages}
\alias{db_get_portal_messages}
\title{Retrieve active portal messages}
\usage{
db_get_portal_messages()
}
\value{
A data frame of portal messages.
}
\description{
Returns active portal messages for the current application, including
global (catch‑all) messages.
}
