% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-record_download.R
\name{db_record_download}
\alias{db_record_download}
\title{Record a File Download Event}
\usage{
db_record_download(
  user = "Guest",
  page_name,
  file_name,
  db_execute = utils_db_execute
)
}
\arguments{
\item{user}{Character. Active Directory username. Default is "Guest".}

\item{page_name}{Character. Page where the download occurred.}

\item{file_name}{Character. Name of the file downloaded.}

\item{db_execute}{Function used to execute SQL (dependency injection).}
}
\value{
Invisibly returns NULL.
}
\description{
Logs a user's download action into the tools_analytics table.
}
