% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_add_event}
\alias{db_ru_add_event}
\title{Add a Point-in-Time Interaction Log}
\usage{
db_ru_add_event(
  event_type_id,
  entity_id,
  entity_type,
  event_date,
  is_completed = 1,
  summary_notes = NULL,
  user_id,
  db_get_query = utils_db_get_query
)
}
\description{
Add a Point-in-Time Interaction Log
}
