% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_add_event_action}
\alias{db_ru_add_event_action}
\title{Add a Pre-Configured Custom Field Rule Blueprint Definition}
\usage{
db_ru_add_event_action(
  event_type_id,
  action_name,
  description = "",
  rule_type = "Character",
  is_required = 0,
  ruesv_id = 0,
  user_id = NULL
)
}
\description{
Add a Pre-Configured Custom Field Rule Blueprint Definition
}
