% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_add_event_sub_variety}
\alias{db_ru_add_event_sub_variety}
\title{Add a New Event Sub-Variety Entry}
\usage{
db_ru_add_event_sub_variety(ruevt_id, name, description = NULL, user_id = NULL)
}
\arguments{
\item{ruevt_id}{Integer scalar. Parent event type ID.}

\item{name}{Character scalar. Name of the sub-cohort/variety.}

\item{description}{Character scalar. Context notes.}

\item{user_id}{Character scalar. Active auditor logging the row.}
}
\description{
Add a New Event Sub-Variety Entry
}
