% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_add_event_sub_variety}
\alias{db_ru_add_event_sub_variety}
\title{Add a New Event Sub-Variety Entry}
\usage{
db_ru_add_event_sub_variety(ruevt_id, name, description = NULL, user_id = NULL)
}
\description{
Add a New Event Sub-Variety Entry
}
