% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_get_event_action_responses}
\alias{db_ru_get_event_action_responses}
\title{Get Collected Dynamic Form Action Response Records}
\usage{
db_ru_get_event_action_responses(event_id, db_get_query = utils_db_get_query)
}
\description{
Pulls configured fields paired alongside existing values logged for an event iteration.
}
