% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_get_event_actions}
\alias{db_ru_get_event_actions}
\title{Get Pre-Configured Custom Event Fields Blueprint Definitions}
\usage{
db_ru_get_event_actions(
  ruevt_id = NULL,
  rueva_id = NULL,
  db_get_query = utils_db_get_query
)
}
\description{
Get Pre-Configured Custom Event Fields Blueprint Definitions
}
