% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_get_event_master_record}
\alias{db_ru_get_event_master_record}
\title{Retrieve Master Profile Metadata for a Point-in-Time Event}
\usage{
db_ru_get_event_master_record(event_master_id)
}
\arguments{
\item{event_master_id}{Integer. The primary key ID of the target event master record.}
}
\value{
A data.frame containing columns \code{ruevm_id}, \code{ruevm_name}, \code{ruevm_active}, \code{created_date}, and \code{created_by}.
}
\description{
Pulls record parameters for a specific primary event container tracker row.
}
