% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_get_event_sub_varieties}
\alias{db_ru_get_event_sub_varieties}
\title{Fetch Registered Sub-Varieties for an Event Type}
\usage{
db_ru_get_event_sub_varieties(ruevt_id = NULL)
}
\arguments{
\item{ruevt_id}{Integer scalar. Filter by parent type (optional).}
}
\description{
Fetch Registered Sub-Varieties for an Event Type
}
