% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_event_helpers.R
\name{db_ru_update_event}
\alias{db_ru_update_event}
\title{Update an Operational Event Record}
\usage{
db_ru_update_event(
  ruev_id,
  event_date,
  is_completed,
  summary_notes = NULL,
  user_id
)
}
\description{
Update an Operational Event Record
}
