% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-user_roles.R
\name{db_update_user_role}
\alias{db_update_user_role}
\title{Update or Insert Application User Role Assignment}
\usage{
db_update_user_role(user_id, role_id, app_id, assigned_by)
}
\arguments{
\item{user_id}{Integer scalar. Target user ID.}

\item{role_id}{Integer scalar. New role ID.}

\item{app_id}{Integer scalar. Application ID.}

\item{assigned_by}{Integer scalar. Admin user ID making the change.}
}
\value{
Integer scalar. Rows affected.
}
\description{
Update or Insert Application User Role Assignment
}
