% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-portal_user_admin_helpers.R
\name{db_update_user_role}
\alias{db_update_user_role}
\title{Insert or update user role for an app}
\usage{
db_update_user_role(conn, user_id, role_id, app_id, assigned_by)
}
\description{
Insert or update user role for an app
}
