% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-portal_user_admin_helpers.R
\name{db_update_user_role}
\alias{db_update_user_role}
\title{Update user role}
\usage{
db_update_user_role(
  user_id,
  role_id,
  app_id,
  assigned_by,
  db_execute = utils_db_execute
)
}
\description{
Update user role
}
