% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_allocations.R
\name{deactivate_ru_lead_allocation}
\alias{deactivate_ru_lead_allocation}
\title{Soft delete / deactivate a lead allocation record}
\usage{
deactivate_ru_lead_allocation(
  rula_id,
  user_id_edited = Sys.getenv("USERNAME", "SYSTEM")
)
}
\arguments{
\item{rula_id}{Integer. The primary key ID to deactivate.}

\item{user_id_edited}{Character. Username or audit identifier.}
}
\value{
Integer number of affected rows, or NULL on error.
}
\description{
Soft delete / deactivate a lead allocation record
}
