% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-ru_lead_schools.R
\name{deactivate_ru_lead_school}
\alias{deactivate_ru_lead_school}
\title{Soft delete / deactivate a lead school record}
\usage{
deactivate_ru_lead_school(
  ruhl_id,
  user_id_edited = Sys.getenv("USERNAME", "SYSTEM")
)
}
\arguments{
\item{ruhl_id}{Integer. The primary key ID to deactivate.}

\item{user_id_edited}{Character. Username or audit identifier.}
}
\value{
Integer number of affected rows, or NULL on error.
}
\description{
Soft delete / deactivate a lead school record
}
