% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_config.R
\name{get_config}
\alias{get_config}
\title{Load portal configuration from YAML}
\usage{
get_config(path = "./config.yml")
}
\arguments{
\item{path}{Path to config YAML file.}
}
\value{
A structured configuration list.
}
\description{
Reads a YAML config file and returns a validated configuration list.
}
