% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-user_context.R
\name{get_user_id}
\alias{get_user_id}
\title{Get user_id from username}
\usage{
get_user_id(conn, username)
}
\arguments{
\item{conn}{A DBI connection from sql_manager("dit")}

\item{username}{A character string (e.g., "bsmith7")}
}
\value{
Integer user_id, or NA_integer_ if not found
}
\description{
Looks up a user's numeric user_id in the AIDT users table.
}
