% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url.R
\name{gias_trust_url}
\alias{gias_trust_url}
\title{Generate GIAS Trust URL}
\usage{
gias_trust_url(trust_ref)
}
\arguments{
\item{trust_ref}{A character string representing the trust reference number.}
}
\value{
A character string containing the full URL to the GIAS trust details page, or \code{NULL} if the reference is invalid.
}
\description{
Constructs a URL to the GIAS page for a given trust reference number.
Cleans the input by removing leading "tr" and zeroes.
}
\examples{
gias_trust_url("TR01234")
gias_trust_url("0001234")
gias_trust_url("") # returns NULL
}
