% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shiny_link.R
\name{make_shiny_links}
\alias{make_shiny_links}
\title{Create Multiple Shiny HTML Links}
\usage{
make_shiny_links(urls)
}
\arguments{
\item{urls}{A named character vector where names are link texts and values are URLs.}
}
\value{
A named list of Shiny HTML anchor tags.
}
\description{
Generates a named list of Shiny HTML anchor tags from a named vector of URLs.
}
\examples{
make_shiny_links(c("DfE" = "https://www.gov.uk/government/organisations/department-for-education"))
}
