% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-quality_add_log.R
\name{quality_add_log}
\alias{quality_add_log}
\title{Insert a New Quality Check Log Entry}
\usage{
quality_add_log(
  quality_check_id,
  live_issues = 0,
  new_issues = 0,
  closed_issues = 0,
  db_execute = utils_db_execute
)
}
\arguments{
\item{quality_check_id}{Integer scalar.}

\item{live_issues}{Integer scalar or NULL/NA.}

\item{new_issues}{Integer scalar or NULL/NA.}

\item{closed_issues}{Integer scalar or NULL/NA.}

\item{db_execute}{Function used to execute SQL (dependency injection).}
}
\value{
Invisibly returns number of rows affected.
}
\description{
Logs the execution of a quality check, recording counts of live, new,
and closed issues.
}
