% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-ru_bi.R
\name{ru_portal_health_server}
\alias{ru_portal_health_server}
\title{RISE Universal Hubs Portal Health Server}
\usage{
ru_portal_health_server(
  id,
  con,
  schema = utils_resolve_schema("db_schema_01r")
)
}
\arguments{
\item{id}{Character module namespace ID.}

\item{con}{DBI SQL Server connection or reactive returning a connection.}

\item{schema}{Character target RISE schema (e.g., "\link{01_RISE}" or "\link{01_RISE_b}").}
}
\description{
RISE Universal Hubs Portal Health Server
}
