% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-school_helpers.R
\name{school_render_overview}
\alias{school_render_overview}
\title{Render School Overview Panel}
\usage{
school_render_overview(urn, id = paste0("school_overview_", urn))
}
\description{
Render School Overview Panel
}
