% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url.R
\name{scp_sc_url}
\alias{scp_sc_url}
\title{Generate Sig Change URL for the sig change portal}
\usage{
scp_sc_url(sigchange_id)
}
\arguments{
\item{sigchange_id}{A character string representing the sigchange_id of a sig change.}
}
\value{
A character string containing the full sig change id to the sc portal, or \code{NULL} if the sc id is invalid.
}
\description{
Constructs a Sig Change URL to our internal portal for a given sigchange_id
}
\examples{
scp_sc_url("1234")
scp_sc_url(NULL) # returns NULL
}
