% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-portal_user_admin.R
\name{server_portal_user_admin}
\alias{server_portal_user_admin}
\title{User/Role Management Server}
\usage{
server_portal_user_admin(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
User/Role Management Server
}
