% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-quality_helpers.R
\name{server_quality_wrapper}
\alias{server_quality_wrapper}
\title{Quality Table Module Server}
\usage{
server_quality_wrapper(id, app_id)
}
\arguments{
\item{id}{Character scalar. Shiny module ID.}

\item{app_id}{Character scalar. The application identifier.}
}
\description{
Quality Table Module Server
}
