% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url.R
\name{slic_urn_url}
\alias{slic_urn_url}
\title{Generate SLIC URL for the SLIC portal}
\usage{
slic_urn_url(urn)
}
\arguments{
\item{urn}{A character vector (or coercible to character) of URNs.}
}
\value{
A character vector containing the full link(s) to the SLIC portal.
Invalid/missing inputs yield NA_character_.
}
\description{
Constructs a SLIC URL to our internal portal for a given URN (vectorised).
}
\examples{
slic_urn_url("1234")
slic_urn_url(c("1234", NA, ""))
}
