% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url.R
\name{slic_urn_url}
\alias{slic_urn_url}
\title{Generate SLIC URL for the SLIC portal}
\usage{
slic_urn_url(urn)
}
\arguments{
\item{sigchange_id}{A character string representing the urn of a school.}
}
\value{
A character string containing the full link to the slic portal, or \code{NULL} if the urn is invalid.
}
\description{
Constructs a SLIC URL to our internal portal for a given urn
}
\examples{
slic_urn_url("1234")
slic_urn_url(NULL) # returns NULL
}
