% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sql_manager.R
\name{sql_manager}
\alias{sql_manager}
\title{Create a SQL Server Connection Using a Named Service Configuration}
\usage{
sql_manager(service, config_file = "./config.yml")
}
\arguments{
\item{service}{Character name of the config section (e.g. "prod", "dev").}

\item{config_file}{Path to the config YAML file.}
}
\value{
A DBI connection object.
}
\description{
Creates an ODBC connection using credentials stored in a config file.
}
