% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-portal_user_admin.R
\name{ui_portal_user_admin}
\alias{ui_portal_user_admin}
\title{User/Role Management UI}
\usage{
ui_portal_user_admin(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
User/Role Management UI
}
