% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-portal_user_admin_modal.R
\name{ui_role_edit_modal}
\alias{ui_role_edit_modal}
\title{Modal dialog for editing a user's role}
\usage{
ui_role_edit_modal(ns, selected, conn)
}
\description{
Modal dialog for editing a user's role
}
\keyword{internal}
