% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dbi.R
\name{utils_db_execute}
\alias{utils_db_execute}
\title{Execute a database command}
\usage{
utils_db_execute(conn, query)
}
\description{
Thin wrapper around DBI::dbExecute().
Required to safely unit-test DB code.
}
