% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dbi.R
\name{utils_db_get_query}
\alias{utils_db_get_query}
\title{Execute a database query and return results}
\usage{
utils_db_get_query(conn, query)
}
\description{
Thin wrapper around DBI::dbGetQuery().
Required to safely unit-test DB code.
}
