% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-environment.R
\name{utils_resolve_env_id}
\alias{utils_resolve_env_id}
\title{Resolve Environment Context Name to Integer Key}
\usage{
utils_resolve_env_id()
}
\value{
Integer scalar matching environments_config map (1 = live, 2 = test, 3 = beta, 4 = dev)
}
\description{
Evaluates the application host URL strings to determine the active running environment.
}
\keyword{internal}
