% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-warning_notice_scrape.R
\name{wn_web_publish_sql}
\alias{wn_web_publish_sql}
\title{Sync Scraped Dataframe into SQL Table}
\usage{
wn_web_publish_sql(post_scrape)
}
\arguments{
\item{post_scrape}{Dataframe. Output dataset stemming from \code{wn_scrape_all_pages}.}
}
\value{
A character message tracking execution operations.
}
\description{
Processes an extracted dataframe against existing records to trigger updates or additions.
}
