% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url.R
\name{wnp_wn_url}
\alias{wnp_wn_url}
\title{Generate Warning Notice URL for the warning notice portal}
\usage{
wnp_wn_url(twn_id)
}
\arguments{
\item{twn_id.}{A character string representing the twn_id of a warning notice.}
}
\value{
A character string containing the full URL to the wn portal, or \code{NULL} if the twn_id is invalid.
}
\description{
Constructs a Warning Notice to our internal portal for a given twn_id.
}
\examples{
wnp_wn_url("1234")
wnp_wn_url(NULL) # returns NULL
}
